Clothing Entity

+++ src/main/java/com/terremotospr/beans/MedicationBean.java
@@ -8,7 +8,7 @@ public class MedicationBean extends BaseResourceBean {
 
     private Long id;
     private String medType;
-    private Long size;
+    private Double size;
     private String dosageForm;
 
     public Long getId() { return id; }
@@ -19,9 +19,9 @@ public class MedicationBean extends BaseResourceBean {
 
     public void setMedType(String medType) { this.medType = medType; }
 
-    public Long getSize() { return size; }
+    public Double getSize() { return size; }
 
-    public void setSize(Long size) { this.size = size; }
+    public void setSize(Double size) { this.size = size; }
 
     public String getDosageForm() { return dosageForm; }
